% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{ant}
\alias{ant}
\title{Ant}
\usage{
ant(dt = 0.001)
}
\arguments{
\item{dt}{Time step}
}
\description{
Generates 'Ant' data from parametric formulas and stores it in a data frame
}
\examples{
ant()
}
\keyword{ant}
