% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_in_flight.R
\name{bird_in_flight}
\alias{bird_in_flight}
\title{Generate bird_in_flight data}
\usage{
bird_in_flight(n = 20000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a bird in flight and stores it in a data frame with (x, y, xend, yend) coordinates
}
\examples{
bird_in_flight()
}
\keyword{bird}
\keyword{flight}
