% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\title{Branch}
\usage{
branch(n = 3000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Branch' data from parametric formulas and stores it in a data frame
}
\examples{
branch()
}
\keyword{leaves}
\keyword{plant}
