% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly1.R
\name{butterfly1}
\alias{butterfly1}
\title{Butterfly 1}
\usage{
butterfly1(n = 40000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Butterfly 1' data from parametric formulas and stores it in a data frame
}
\examples{
butterfly1()
}
\keyword{butterfly}
