% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_curve.R
\name{butterfly_curve}
\alias{butterfly_curve}
\title{Butterfly curve}
\usage{
butterfly_curve(
  n = 1000,
  a = 12,
  b = 12,
  c = 2,
  d = 2,
  e = 4,
  f = 4,
  g = 5,
  h = 5,
  i = 12,
  j = 12
)
}
\arguments{
\item{n}{The number of points to generate}

\item{a}{A parameter}

\item{b}{A parameter}

\item{c}{A parameter}

\item{d}{A parameter}

\item{e}{A parameter}

\item{f}{A parameter}

\item{g}{A parameter}

\item{h}{A parameter}

\item{i}{A parameter}

\item{j}{A parameter}
}
\description{
Generates data for plotting a generalised butterfly curve (Fay, 1989) and stores it in a data frame with (x, y) coordinates. Refer to the \href{https://en.wikipedia.org/wiki/Butterfly_curve_(transcendental)}{Wikipedia article} for details.
}
\examples{
butterfly_curve()
}
\keyword{butterfly}
\keyword{curve}
