% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave_hull.R
\name{concave_hull}
\alias{concave_hull}
\title{Concave hull}
\usage{
concave_hull(points, concavity = 1, length_threshold = 0)
}
\arguments{
\item{points}{A data frame of points.}

\item{concavity}{A value that determines the concavity (1 = detailed shape, Inf = convex hull).}

\item{length_threshold}{Segment lengths less than this threshold are not considered; higher values give simpler shapes.}
}
\description{
Computes the concave hull for a set of points (wrapper for concaveman::concaveman)
}
\examples{
concave_hull()
}
\keyword{concave}
\keyword{hull}
