% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicycloid.R
\name{epicycloid_curve}
\alias{epicycloid_curve}
\title{Epicycloid curve}
\usage{
epicycloid_curve(a = 1, b = 1, delta = 0.1 * pi)
}
\arguments{
\item{a}{Curve parameter}

\item{b}{Curve parameter}

\item{delta}{Timestep}
}
\description{
Generates data for plotting an epicycloid curve and stores it in a data frame with (x, y) coordinates. Refer to the \href{https://en.wikipedia.org/wiki/Epicycloid}{Wikipedia article} for details.
}
\examples{
epicycloid_curve()
}
\keyword{epicycloid}
