% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower7.R
\name{flower7}
\alias{flower7}
\title{Flower 7}
\usage{
flower7(n = 9000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Flower 7' data from parametric formulas and stores it in a data frame (https://www.huffingtonpost.com/entry/flowers_b_9817126.html)
}
\examples{
flower7()
}
\keyword{flower}
