% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal_fern.R
\name{fractal_fern}
\alias{fractal_fern}
\title{Generate fractal fern data}
\usage{
fractal_fern(n = 10000, a = c(0, 0.85, 0.2, -0.15), b = c(0, 0.04, -0.26,
  0.28), c_ = c(0, -0.04, 0.23, 0.26), d = c(0.16, 0.85, 0.22, 0.24),
  e = c(0, 0, 0, 0), f = c(0, 1.6, 1.6, 0.44), p = c(0.01, 0.85, 0.07,
  0.07))
}
\arguments{
\item{n}{The number of points to generate}

\item{a}{Coefficient vector of length 4}

\item{b}{Coefficient vector of length 4}

\item{d}{Coefficient vector of length 4}

\item{e}{Coefficient vector of length 4}

\item{f}{Coefficient vector of length 4}

\item{p}{Probability vector of length 4}

\item{c}{Coefficient vector of length 4}
}
\description{
Generates data for plotting a fractal fern and stores it in a data frame with (x, y) coordinates
}
\examples{
fractal_fern()
}
\keyword{fern}
\keyword{fractal}
