% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonograph.R
\name{harmonograph}
\alias{harmonograph}
\title{Generate harmonograph data}
\usage{
harmonograph(n = 1e+06, A1 = 1, A2 = 1, A3 = 1, A4 = 1, d1 = 0.039,
  d2 = 0.006, d3 = 0, d4 = 0.0045, f1 = 10, f2 = 3, f3 = 1,
  f4 = 2, p1 = 0, p2 = 0, p3 = pi/2, p4 = 0)
}
\arguments{
\item{n}{The number of points to generate}

\item{A1}{A parameter}

\item{A2}{A parameter}

\item{A3}{A parameter}

\item{A4}{A parameter}

\item{d1}{A parameter}

\item{d2}{A parameter}

\item{d3}{A parameter}

\item{d4}{A parameter}

\item{f1}{A parameter}

\item{f2}{A parameter}

\item{f3}{A parameter}

\item{f4}{A parameter}

\item{p1}{A parameter}

\item{p2}{A parameter}

\item{p3}{A parameter}

\item{p4}{A parameter}
}
\description{
Generates data for plotting a harmonograph and stores it in a data frame with (x, y) coordinates. Refer to the \href{https://en.wikipedia.org/wiki/Harmonograph}{Wikipedia article} for details.
}
\examples{
harmonograph()
}
\keyword{harmonograph}
