% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart1.R
\name{heart1}
\alias{heart1}
\title{Generate heart1 data}
\usage{
heart1(n = 1e+06)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a heart and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
heart1()
}
\keyword{heart}
