% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart4.R
\name{heart4}
\alias{heart4}
\title{Generate heart4 data}
\usage{
heart4(n = 1e+06)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a heart and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/mathematical-formulas-for-the-symbol-of-love_b_8848264.html}{here}.
}
\examples{
heart4()
}
\keyword{heart}
