% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_nearest_neighbour_graph.R
\name{k_nearest_neighbour_graph}
\alias{k_nearest_neighbour_graph}
\title{k-nearest neighbour graph}
\usage{
k_nearest_neighbour_graph(points, k = 8)
}
\arguments{
\item{points}{A data frame with x, y coordinates for the points}

\item{k}{Number of neighbours}
}
\description{
Computes a k-nearest neighbour graph for a given set of points. Refer to the \href{https://en.wikipedia.org/wiki/Nearest_neighbor_graph}{Wikipedia article} for details.
}
\examples{
k_nearest_neighbour_graph()
}
\keyword{graph}
\keyword{nearest}
\keyword{neightbour}
