% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{kdtree}
\alias{kdtree}
\title{k-d tree}
\usage{
kdtree(points, minmax = FALSE)
}
\arguments{
\item{points}{A data frame with columns for x and y coordinates, and each point in a row. Refer to the \href{https://en.wikipedia.org/wiki/K-d_tree}{Wikipedia article} for details.}
}
\description{
Computes a k-d tree for a given set of points
}
\examples{
kdtree()
}
\keyword{k-d}
\keyword{tree}
