% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaves.R
\name{leaves}
\alias{leaves}
\title{Leaves}
\usage{
leaves(n = 30000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Leaves' data from parametric formulas and stores it in a data frame
}
\examples{
leaves()
}
\keyword{leaves}
\keyword{plant}
