% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissajous.R
\name{lissajous}
\alias{lissajous}
\title{Generate Lissajous data}
\usage{
lissajous(a = 1, b = 1, A = 1, B = 1, d = 200, delta = 0.01 * pi)
}
\arguments{
\item{a}{<- Parameter}

\item{b}{<- Parameter}

\item{A}{<- Parameter}

\item{B}{<- Parameter}

\item{d}{<- Parameter}

\item{delta}{<- Timestep}

\item{n}{The number of points to generate}
}
\description{
Generates data for plotting Lissajous curves and stores it in a data frame with (x, y) coordinates. Refer to the \href{https://en.wikipedia.org/wiki/Lissajous_curve}{Wikipedia article} for details.
}
\examples{
lissajous()
}
\keyword{lissajous}
