% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_attractor.R
\name{lorenz_attractor}
\alias{lorenz_attractor}
\title{Lorenz attractor}
\usage{
lorenz_attractor(a = 2, b = 0.7, c = 26, x0 = -1, y0 = 1, z0 = -1,
  dt = 0.01, n = 10000)
}
\arguments{
\item{a, b, c}{Parameters}

\item{x0, y0, z0}{Initial point coordinates}

\item{dt}{Time step}

\item{n}{Number of points to generate}
}
\description{
Generates data for plotting a Lorenz attractor and stores it in a data frame. The Rcpp implementation is from the blog post \href{https://fronkonstin.com/2017/11/07/drawing-10-million-points-with-ggplot-clifford-attractors/}{Drawing 10 Million Points With ggplot} by Antonio Sanchez Chinchon.
}
\examples{
lorenz_attractor()
}
\keyword{Lorenz}
\keyword{attractor}
