% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millipede.R
\name{millipede}
\alias{millipede}
\title{Millipede}
\usage{
millipede(dt = 0.001)
}
\arguments{
\item{dt}{Time step}
}
\description{
Generates 'Millipede' data from parametric formulas and stores it in a data frame
}
\examples{
millipede()
}
\keyword{millipede}
