% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mollusc.R
\name{mollusc}
\alias{mollusc}
\title{Mollusc}
\usage{
mollusc(alpha = 80, beta = 40, phi = 55, mu = 10, Omega = 30,
  s_min = -270, s_max = 62, A = 25, a = 12, b = 16, P = 2,
  W_1 = 1, W_2 = 1, N = 1, L = 0, D = 1, theta_start = -100 * 0,
  theta_end = 10 * pi)
}
\arguments{
\item{dt}{Time step}
}
\description{
Generates 'Mollusc' data from parametric formulas and stores it in a data frame
}
\examples{
ant()
}
\keyword{ant}
