% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olive_branch.R
\name{olive_branch}
\alias{olive_branch}
\title{Olive branch}
\usage{
olive_branch(n = 4000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Olive branch' data from parametric formulas and stores it in a data frame
}
\examples{
olive_branch()
}
\keyword{branch}
\keyword{olive}
\keyword{plant}
