% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_branch.R
\name{palm_branch}
\alias{palm_branch}
\title{Generate palm_branch data}
\usage{
palm_branch(n = 12000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a palm branch and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/mathematical-plants_b_9010790.html}{here}.
}
\examples{
palm_branch()
}
\keyword{branch}
\keyword{palm}
\keyword{plant}
