% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{parrot}
\alias{parrot}
\title{Generate parrot data}
\usage{
parrot(n = 2e+05)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a parrot and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/mathematical-birds_b_8876904.html}{here}.
}
\examples{
parrot()
}
\keyword{parrot}
