% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi1}
\alias{pi1}
\title{Pi}
\usage{
pi1(n = 2000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Pi' data from parametric formulas and stores it in a data frame
}
\examples{
pi1()
}
\keyword{pi}
