% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadtree.R
\name{quadtree}
\alias{quadtree}
\title{Quadtree}
\usage{
quadtree(points)
}
\arguments{
\item{points}{A data frame with columns for x and y coordinates, and each point in a row.}
}
\description{
Compute a quadtree for a given set of points. Refer to the \href{https://en.wikipedia.org/wiki/Quadtree}{Wikipedia article} for details.
}
\examples{
quadtree()
}
\keyword{quadtree}
