% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidly_exploring_random_tree.R
\name{rapidly_exploring_random_tree}
\alias{rapidly_exploring_random_tree}
\title{Rapidly-exploring random tree}
\usage{
rapidly_exploring_random_tree(n = 10000, X = 1000, delta = 2.5)
}
\arguments{
\item{n}{Number of points to generate}

\item{X}{Square canvas dimension}

\item{delta}{Maximum edge length}
}
\description{
Generates data for plotting a rapidly exploring random tree (rrt). Refer to the \href{https://en.wikipedia.org/wiki/Rapidly-exploring_random_tree}{Wikipedia article} for details.
}
\examples{
rapidly_exploring_random_tree()
}
\keyword{exploring}
\keyword{random}
\keyword{rapidly}
\keyword{tree}
