% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sand_paint.R
\name{sand_paint}
\alias{sand_paint}
\title{Sand paint}
\usage{
sand_paint(edges, n_grains = 100)
}
\arguments{
\item{n_grains}{Number of grains distributed across each line segment}

\item{points}{A data frame with x, y, xend, yend coordinates for the line segments}
}
\description{
Sand paints a set of line segments
}
\examples{
sand_point()
}
\keyword{paint}
\keyword{sand}
