% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_noise.R
\name{segment_noise}
\alias{segment_noise}
\title{Segment noise}
\usage{
segment_noise(segments, eps, N, nframes)
}
\arguments{
\item{eps}{aximum endpoint perturbation as a percentage of edge length}

\item{N}{Number of edge replications}

\item{nframes}{Number of tween iterations}

\item{segmments}{A data frame with segment endpoint coordinates x, y, xend and yend}
}
\description{
Adds noise to a collection of line segments
}
\examples{
segment_noise()
}
\keyword{noise}
\keyword{segment}
