% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stork.R
\name{stork}
\alias{stork}
\title{Generate stork data}
\usage{
stork(n = 80000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a stork and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
stork()
}
\keyword{stork}
