% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{tessellate}
\alias{tessellate}
\title{Create a tessellatation for a given set of points}
\usage{
tessellate(points)
}
\arguments{
\item{points}{A data frame with point x and y coordinates}
}
\value{
A data frame with segment x, y, xend and yend coorindates
}
\description{
Creates a tessellatation for a given set of points (wrapper for deldir)
}
\details{
A simple wrapper for deldir::deldir()
}
\examples{
data.frame(x = runif(10), y = runif(10)) \%>\% tessellate()
}
