% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{Create a triangulation for a given set of points}
\usage{
triangulate(points)
}
\arguments{
\item{points}{A data frame with point x and y coordinates}
}
\value{
A data frame with segment x, y, xend and yend coorindates
}
\description{
Creates a triangulation for a given set of points (wrapper for deldir)
}
\details{
A simple wrapper for deldir::deldir()
}
\examples{
data.frame(x = runif(10), y = runif(10)) \%>\% triangulate()
}
