% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiszfeld.R
\name{weiszfeld}
\alias{weiszfeld}
\title{Weiszfeld}
\usage{
weiszfeld(points, p0)
}
\arguments{
\item{points}{A data frame with columns for x and y coordinates}

\item{p0}{A vector with the (x, y) co-ordinates of the initial point}
}
\description{
Finds the geometric median for a given set of points using Weiszfeld's algorithm. Refer to the \href{https://en.wikipedia.org/wiki/Geometric_median}{Wikipedia article} for details.
}
\examples{
weiszfeld()
}
\keyword{weiszfeld}
