% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{ant}
\alias{ant}
\title{Generate ant data}
\usage{
ant(n = 3000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting an ant and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
ant()
}
\keyword{ant}
