% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\title{Generate branch data}
\usage{
branch(n = 3000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a branch and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/mathematical-plants_b_9010790.html}{here}.
}
\examples{
branch()
}
\keyword{branch}
