% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly1.R
\name{butterfly1}
\alias{butterfly1}
\title{Generate butterfly1 data}
\usage{
butterfly1(n = 40000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a butterfly and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/using-mathematical-formul_b_9313484.html}{here}.
}
\examples{
butterfly1()
}
\keyword{butterfly}
