% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly1.R
\name{butterfly1}
\alias{butterfly1}
\title{Generate butterfly1 data}
\usage{
butterfly1(n = 40000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a butterfly and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
butterfly1()
}
\keyword{butterfly}
