% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly2.R
\name{butterfly2}
\alias{butterfly2}
\title{Generate butterfly2 data}
\usage{
butterfly2(n = 40000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a butterfly and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
butterfly2()
}
\keyword{butterfly}
