% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly3.R
\name{butterfly3}
\alias{butterfly3}
\title{Generate butterfly3 data}
\usage{
butterfly3(n = 40000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a butterfly and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
butterfly3()
}
\keyword{butterfly}
