% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower7.R
\name{flower7}
\alias{flower7}
\title{Generate flower7 data}
\usage{
flower7(n = 9000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a flower and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
flower7()
}
\keyword{flower}
