% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart2.R
\name{heart2}
\alias{heart2}
\title{Generate heart2 data}
\usage{
heart2(n = 1e+06)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a heart and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
heart2()
}
\keyword{heart}
