% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart3.R
\name{heart3}
\alias{heart3}
\title{Generate heart3 data}
\usage{
heart3(n = 1e+06)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a heart and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
heart3()
}
\keyword{heart}
