% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaves.R
\name{leaves}
\alias{leaves}
\title{Generate leaves data}
\usage{
leaves(n = 30000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting leaves and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
leaves()
}
\keyword{leaves}
\keyword{plant}
