% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpie.R
\name{magpie}
\alias{magpie}
\title{Generate magpie data}
\usage{
magpie(n = 2e+05)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a magpie and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
magpie()
}
\keyword{magpie}
