% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millipede.R
\name{millipede}
\alias{millipede}
\title{Generate millipede data}
\usage{
millipede(n = 3000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a millipede and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/ant-spider-and-millipede-_b_8583338.html}{here}.
}
\examples{
millipede()
}
\keyword{millipede}
