% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mollusc.R
\name{mollusc}
\alias{mollusc}
\title{Generate mollusc data}
\usage{
mollusc(n = 1000, alpha = 80, beta = 40, phi = 55, mu = 10,
  Omega = 30, s_min = -270, s_max = 62, A = 25, a = 12, b = 16,
  P = 2, W_1 = 1, W_2 = 1, N = 1, L = 0, D = 1, theta_start = 0,
  theta_end = 10 * pi)
}
\arguments{
\item{n}{Number of points to generate}

\item{alpha}{Equiangular angle of spiral}

\item{beta}{Angle between z-axis and line from aperture local origin to xyz origin}

\item{phi}{Tilt of ellipse major axis from horizontal plane}

\item{mu}{Amount of "leaning over" of aperture}

\item{Omega}{Amount of azimuthal rotation of aperture}

\item{s_min}{Angle at which aperture-generating curve begins}

\item{s_max}{Angle at which aperture-generating curve ends}

\item{A}{Distance from main origin to local origin of aperture at theta=0}

\item{a}{Major radius of ellipse at theta=0}

\item{b}{Minor radius of ellipse at theta=0}

\item{P}{Position of nodule in terms of the angle, s}

\item{W_1}{Width of nodule in s-direction}

\item{W_2}{Width of nodule in theta-direction}

\item{N}{Number of nodules per whorl}

\item{L}{Height of nodule at theta=0}

\item{D}{Sense of coiling; 1=dextral, -1=sinistral}

\item{theta_start}{Required for shells such as Dentalium or Diodora, which grow at one end while dissolving at the other. In all other cases set to -Inf}

\item{theta_end}{Not important for self-similar shells, except very small or large values can cause computational problems}
}
\description{
Generates data for plotting mollusc shells and stores it in a data frame with (x, y, z) coordinates. The shell model is described in the paper "Models for mollusc shell shape" by M.B. Cortie (1989).
}
\examples{
mollusc()
}
\keyword{mollusc}
