% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olive_branch.R
\name{olive_branch}
\alias{olive_branch}
\title{Generate olive_branch data}
\usage{
olive_branch(n = 4000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting an olive branch and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
olive_branch()
}
\keyword{branch}
\keyword{olive}
\keyword{plant}
