% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palm_branch.R
\name{palm_branch}
\alias{palm_branch}
\title{Olive branch}
\usage{
palm_branch(n = 12000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates 'Palm branch' data from parametric formulas and stores it in a data frame
}
\examples{
palm_branch()
}
\keyword{branch}
\keyword{palm}
\keyword{plant}
