% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parrot.R
\name{parrot}
\alias{parrot}
\title{Generate parrot data}
\usage{
parrot(n = 2e+05)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a parrot and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
parrot()
}
\keyword{parrot}
