% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi1}
\alias{pi1}
\title{Generate pi data}
\usage{
pi1(n = 2000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting the pi symbol and stores it in a data frame with (x, y) coordinates and radius r
}
\details{
The equations were discovered by the mathematical artist \href{https://mathematics.culturalspot.org/home}{Hamid Naderi Yeganeh}. \cr

Hamid's equations can be found \href{https://www.huffingtonpost.com/entry/a-mathematical-artwork-fo_b_9450174.html}{here}.
}
\examples{
pi1()
}
\keyword{pi}
