% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Spider}
\usage{
spider(dt = 0.001)
}
\arguments{
\item{dt}{Time step}
}
\description{
Generates 'Spider' data from parametric formulas and stores it in a data frame
}
\examples{
spider()
}
\keyword{spider}
