% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Generate spider data}
\usage{
spider(n = 3000)
}
\arguments{
\item{n}{The number of points to generate}
}
\description{
Generates data for plotting a spider and stores it in a data frame with (x, y) coordinates and radius r
}
\examples{
spider()
}
\keyword{spider}
